Fix keyword search hightlighting
The simple gsub search highlighting I had implemented was breaking links when
the keywork would appear in the url. This change fixes that, and should fix all
similar issues.

Insead of preform a gsub on the unprocessed source, it waits until the document
has been parsed, and then walks through all the text elements in the document
and swaps them with html, with the highlight span embedded.